requirements.txt
torch
torchvision
transformers
pillow
matplotlib
requests
duckduckgo-search